fix : 아주 큰 수의 합산이 가능하도록, 합산 결과 타입을 BigInteger 로 변경
- 모든 요소를 BigInteger 로 처리하면 메모리 상 불리하므로, 요소 하나하나는 Long 타입으로 처리

+++ src/main/java/calculator/calculator/CalculatorController.java
@@ -1,6 +1,7 @@
 package calculator.calculator;
 
 import camp.nextstep.edu.missionutils.Console;
+import java.math.BigInteger;
 
 public class CalculatorController {
     private static final String INITIAL_COMMENT = "덧셈할 문자열을 입력해 주세요.";
@@ -11,7 +12,7 @@ public class CalculatorController {
         System.out.println(INITIAL_COMMENT);
         Calculator calculator = new Calculator();
         String input = Console.readLine();
-        int result = calculator.calculate(input);
+        BigInteger result = calculator.calculate(input);
         System.out.println(RESULT + result);
         Console.close();
     }
